# Conflicts:
#	common/meta/authority.go
#	common/meta/authority_test.go
#	common/meta/identityManager.go
#	common/meta/identityUtils.go
#	common/meta/identityUtils_test.go
#	common/meta/newMatryoshkaHash_test.go
#	engine/NetStart.go
#	state/saveAndRestore.go

+++ state/identityUtils.go
@@ -98,7 +98,7 @@ func CheckTimestamp(time []byte, timeEntered int64) bool {
 	}
 }
 
-func statusIsFedOrAudit(status int) bool {
+func statusIsFedOrAudit(status uint8) bool {
 	if status == constants.IDENTITY_FEDERATED_SERVER ||
 		status == constants.IDENTITY_AUDIT_SERVER ||
 		status == constants.IDENTITY_PENDING_FEDERATED_SERVER ||
